+++ hawk/public/stylesheets/hawk.css
@@ -11,6 +11,7 @@
 @import url('yard/error-page.css');
 @import url('yard/article.css');
 @import url('yard/unsubscribe.css');
+@import url('yard/invite.css');
 
 /**
  * Garage
