+++ hawk/public/stylesheets/yard/invite.css
@@ -1,0 +1,43 @@
+.invite-page {
+
+    padding: 60px 70px;
+    color: #fff;
+
+    &__title {
+        margin: 70px 0 0;
+        font-size: 40px;
+
+        a {
+            color: inherit;
+            text-decoration: none;
+        }
+    }
+
+    &__caption {
+        margin: 8px 0;
+        color: color(#fff alpha(50%));
+        font-size: 20px;
+    }
+
+    &__header {
+        margin: 25px 0 15px;
+        font-size: 27px;
+        font-weight: 400;
+     }
+
+    &__description {
+        max-width: 380px;
+        font-size: 19px;
+        line-height: 1.5em;
+    }
+
+    &__garage-link {
+
+        font-size: 20px;
+        color: var(--linkOnDark);
+
+        svg {
+            margin-left: 0.6em;
+        }
+    }
+}
