--- hawk/views/error.twig
@@ -1,7 +1,0 @@
-{% extends 'layout.twig' %}
-
-{% block body %}
-  <h1>{{message}}</h1>
-  <h2>{{error.status}}</h2>
-  <pre>{{error.stack}}</pre>
-{% endblock %}
