--- hawk/views/user.twig
@@ -1,5 +1,0 @@
-{% extends 'layout.twig' %}
-
-{% block body %}
-  <h1>Welcome, {{user}}</h1>
-{% endblock %}
